Initiaial PG15 (#1108)
Initial PG15 version

Contributors

Panagiotis Foliadis <pfoliadis@hotmail.com>
Matheus Farias <matheusfarias519@gmail.com>
Mohamed Mokhtar <m.mokhtar@outlook.it>
Hannan Aamir <hannanaamir2000@gmail.com>

---------

+++ src/backend/utils/adt/agtype.c
@@ -5551,7 +5551,7 @@ Datum age_tointeger(PG_FUNCTION_ARGS)
                  * return null.
                  */
                 if (!is_valid || isnan(f) || isinf(f) ||
-                    f < PG_INT64_MIN || f > (double)PG_INT64_MAX)
+                    f < (float8)PG_INT64_MIN || f > (float8)PG_INT64_MAX)
                 {
                     PG_RETURN_NULL();
                 }
